fix(tauri): resolve E0599/E0603/E0609/E0277/E0308 visibility and field-name errors
E0603 (private module/function used in public path):
- video/mod.rs:8 'mod processor' -> 'pub mod processor' so lib.rs:35
  'pub use video::processor::VideoProcessor' resolves. The lib.rs
  re-export was already pub, the underlying module just lacked pub.
- director/states.rs:52 'fn unix_timestamp' -> 'pub(crate) fn
  unix_timestamp' so director/commands.rs can call it.

E0599 (method not found on type):
- commands/ai/detection.rs:23 'detector.get_highlights' is wrong —
  the refactor moved it to a free function in highlight/combiner.rs.
  Call crate::highlight::combiner::get_highlights directly. (The
  HighlightDetector struct remains used by the parameter check above.)

E0609 (no field X on struct Y):
- postprocess.rs:40 'm.start_time/m.end_time' -> 'm.start/m.end' —
  AutonomousOverlayMarker in types.rs:80 uses 'start: f64, end: f64'.
- autonomous_cut/mod.rs:104 'input.subtitle' -> 'input.subtitles'
  and use the first item's text, because AutonomousRenderInput
  defines 'subtitles: Option<Vec<AutonomousSubtitle>>' (line 96), no
  singular 'subtitle' field.

E0277 (f32 cannot be made by summing i16):
- highlight/energy.rs:120 's * s' overflows in i16, then can't sum
  to f32. Cast each sample: '(s as f32) * (s as f32)'. This is also
  the correct RMS-style normalization.
- Re-add the 'pcm_data: &[i16]' parameter that an earlier patch
  accidentally removed (it was used 3 lines below).

E0308 (mismatched type):
- subtitle/whisper.rs:46 'lang_arg' is &str but the surrounding
  [..].concat() produces a Vec<String>. Cast: lang_arg.to_string().

Plus a tiny side fix:
- director/types.rs ScriptStylePreset needs Default for
  state_ops::parse_style's '.unwrap_or_default()' on line 19. Add
  #[derive(Default)] and mark Humorous as #[default].

+++ src-tauri/src/highlight/energy.rs
@@ -117,7 +117,7 @@ pub fn compute_energy_profile(
 
     for i in (0..pcm_data.len().saturating_sub(window_samples)).step_by(hop_size) {
         let window = &pcm_data[i..i + window_samples];
-        let energy: f32 = window.iter().map(|&s| s * s).sum::<f32>() / window_samples as f32;
+        let energy: f32 = window.iter().map(|&s| (s as f32) * (s as f32)).sum::<f32>() / window_samples as f32;
         let time_ms = (i as f32 * 1000.0 / sample_rate as f32) as u64;
         energies.push(energy);
         timestamps.push(time_ms);
